# swagger.yaml
openapi: 3.0.3
info:
  title: Telegram Clone API
  description: |
    A comprehensive API for a Telegram clone with Signal Protocol encryption, real-time messaging, and advanced group features.
    
    ## Features
    - 🔐 **End-to-End Encryption** using Signal Protocol
    - 💬 **Real-time Messaging** with WebSocket support
    - 👥 **Advanced Group Types** (Channels, Supergroups, Forums)
    - 📁 **File Upload** and media sharing
    - 🔒 **JWT Authentication** with refresh tokens
    - 📱 **Cross-platform** support
    
    ## Authentication
    Most endpoints require authentication using JWT Bearer tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-access-token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to 100 requests per 15 minutes per IP address.
    
    ## WebSocket
    Real-time messaging is available via WebSocket connection at `ws://localhost:3000`
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@telegramclone.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.telegramclone.com/api/v1
    description: Production server
...
paths:
...
  /users/profile:
    get:
      tags:
        - Users
      summary: Get user profile
      description: Get current user's profile information
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
    put:
      tags:
        - Users
      summary: Update user profile
      description: Update current user's profile information
      operationId: updateProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
  /users/search:
    get:
      tags:
        - Users
      summary: Search users
      description: Search for users by username, email, or phone number
      operationId: searchUsers
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          description: Search query
          required: true
          schema:
            type: string
            minLength: 2
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSearchResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
  /users/{userId}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Get user information by user ID
      operationId: getUserById
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
        username:
          type: string
          description: Username
        email:
          type: string
          format: email
          description: Email address
        phoneNumber:
          type: string
          description: Phone number
        firstName:
          type: string
          description: First name
        lastName:
          type: string
          description: Last name
        bio:
          type: string
          description: User bio
        profilePicture:
          type: string
          format: uri
          description: Profile picture URL
        isOnline:
          type: boolean
          description: Online status
        lastSeen:
          type: string
          format: date-time
          description: Last seen timestamp
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 50
          description: First name
        lastName:
          type: string
          maxLength: 50
          description: Last name
        bio:
          type: string
          maxLength: 500
          description: User bio
        profilePicture:
          type: string
          format: uri
          description: Profile picture URL
    UserSearchResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
          description: Total number of users found
        limit:
          type: integer
          description: Number of users returned
        offset:
          type: integer
          description: Number of users skipped

    # Chat Schemas
...
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        message:
          type: string
          example: "Invalid input data"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field name
              message:
                type: string
                description: Error message
        code:
          type: string
          example: "VALIDATION_ERROR"
    AuthError:
      type: object
      properties:
        error:
          type: string
          example: "Unauthorized"
        message:
          type: string
          example: "Invalid or missing authentication token"
        code:
          type: string
          example: "AUTH_ERROR"
...
    NotFoundError:
      type: object
      properties:
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "Resource not found"
        code:
          type: string
          example: "NOT_FOUND_ERROR"
...